fs: return (int)0 from mmap() in kernel on success
We assume mmap() in kernel returns the same address as the one provided
to mmap(). We should not make the interface look like returning
a different address from the one provided to mmap() is an option.

+++ kernel/fs/fs.c
@@ -194,9 +194,8 @@ ssize_t file_description_write(file_description* desc, const void* buffer,
     return inode->fops->write(desc, buffer, count);
 }
 
-uintptr_t file_description_mmap(file_description* desc, uintptr_t addr,
-                                size_t length, off_t offset,
-                                uint16_t page_flags) {
+int file_description_mmap(file_description* desc, uintptr_t addr, size_t length,
+                          off_t offset, uint16_t page_flags) {
     struct inode* inode = desc->inode;
     if (!inode->fops->mmap)
         return -ENODEV;
