+++ kernel/growable_buf.h
@@ -22,9 +22,8 @@ NODISCARD ssize_t growable_buf_append(growable_buf*, const void* bytes,
 
 NODISCARD int growable_buf_truncate(growable_buf*, off_t length);
 
-NODISCARD uintptr_t growable_buf_mmap(growable_buf*, uintptr_t addr,
-                                      size_t length, off_t offset,
-                                      uint16_t page_flags);
+NODISCARD int growable_buf_mmap(growable_buf*, uintptr_t addr, size_t length,
+                                off_t offset, uint16_t page_flags);
 
 int growable_buf_printf(growable_buf*, const char* format, ...);
 int growable_buf_vsprintf(growable_buf*, const char* format, va_list args);
